Create htm.code
A software developer who has built over 20 websites decided to build a reputable hosting platform. The developer makes use of 3 different CSS files the first contain the font family and the second contains the background and the third referencing the body and paragraph. The developer uploads all necessary files on the server and renders high quality web faced file, inside the developer server consist of HTML 5 file and start with a doctype and uses 10 different tags by default as a developer kindly render this  files correctly.